+++ Engine.sln
@@ -17,7 +17,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "XRSceneTest", "XRSceneTest\
 		{F8BF6CF0-A168-4EE7-A71C-9462A21B2386} = {F8BF6CF0-A168-4EE7-A71C-9462A21B2386}
 	EndProjectSection
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "XRRenderEngineGL", "XRRenderEngineGL\XRRenderEngineGL.vcxproj", "{66E3C0F6-F9AB-4C36-AE40-5CA2B4F18E16}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "XRRHIOpenGL", "XRRHIOpenGL\XRRHIOpenGL.vcxproj", "{66E3C0F6-F9AB-4C36-AE40-5CA2B4F18E16}"
 	ProjectSection(ProjectDependencies) = postProject
 		{F8BF6CF0-A168-4EE7-A71C-9462A21B2386} = {F8BF6CF0-A168-4EE7-A71C-9462A21B2386}
 	EndProjectSection
