+++ Engine.xcodeproj/xcshareddata/xcschemes/XRRHIOpenGL.xcscheme
@@ -15,8 +15,8 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "E8C1BA8122B63B8A006EDA1E"
-               BuildableName = "libXRRenderEngineGL.dylib"
-               BlueprintName = "XRRenderEngineGL"
+               BuildableName = "libXRRHIOpenGL.dylib"
+               BlueprintName = "XRRHIOpenGL"
                ReferencedContainer = "container:Engine.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
@@ -44,8 +44,8 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "E8C1BA8122B63B8A006EDA1E"
-            BuildableName = "libXRRenderEngineGL.dylib"
-            BlueprintName = "XRRenderEngineGL"
+            BuildableName = "libXRRHIOpenGL.dylib"
+            BlueprintName = "XRRHIOpenGL"
             ReferencedContainer = "container:Engine.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
@@ -60,8 +60,8 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "E8C1BA8122B63B8A006EDA1E"
-            BuildableName = "libXRRenderEngineGL.dylib"
-            BlueprintName = "XRRenderEngineGL"
+            BuildableName = "libXRRHIOpenGL.dylib"
+            BlueprintName = "XRRHIOpenGL"
             ReferencedContainer = "container:Engine.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
