+++ XRFrameworkBase/XRBuffer.h
@@ -6,7 +6,7 @@ struct XRBufferCreateInfo;
 class XRBuffer;
 class XRBufferRHI;
 
-#ifdef XRRENDERENGINEGL_EXPORTS
+#ifdef XRRHIOPENGL_EXPORTS
 XRRenderExport XRBuffer* xrCreateBuffer(XRBufferCreateInfo const* createInfo);
 #else
 extern XRBuffer* (*xrCreateBuffer)(XRBufferCreateInfo const* createInfo);
@@ -24,7 +24,7 @@ public:
 
 class XRBaseExport XRBuffer
 {
-#ifdef XRRENDERENGINEGL_EXPORTS
+#ifdef XRRHIOPENGL_EXPORTS
 	/** @brief	Factory function */
 	friend XRBuffer* xrCreateBuffer(XRBufferCreateInfo const* createInfo);
 #endif
