+++ XRRHIOpenGL/XRRenderGroupGL.cpp
@@ -1,6 +1,6 @@
 //
 //  XRRenderGroupGL.cpp
-//  XRRenderEngineGL
+//  XRRHIOpenGL
 //
 //  Created by Jiman Jeong on 2020/10/01.
 //  Copyright © 2020 XRManet. All rights reserved.
