+++ XRRHIOpenGL/XRRenderGroupGL.h
@@ -1,6 +1,6 @@
 ﻿//
 //  XRRenderGroupGL.h
-//  XRRenderEngineGL
+//  XRRHIOpenGL
 //
 //  Created by Jiman Jeong on 2020/10/01.
 //  Copyright © 2020 XRManet. All rights reserved.
