+++ XRRHIOpenGL/stdafx.cpp
